ThemeUtils merged to XamlUtils
