+++ Template10 (Library)/Template10 (Library).csproj
@@ -193,7 +193,6 @@
     <Compile Include="Utils\ResourceUtils.cs" />
     <Compile Include="Utils\TaskUtils.cs" />
     <Compile Include="Utils\Template10Utils.cs" />
-    <Compile Include="Utils\ThemeUtils.cs" />
     <Compile Include="Utils\XamlUtils.cs" />
     <Compile Include="Controls\HamburgerMenu.xaml.cs">
       <DependentUpon>HamburgerMenu.xaml</DependentUpon>
