+++ Template10 (Library)/Utils/XamlUtils.cs
@@ -57,5 +57,31 @@ namespace Template10.Utils
             }
             return list;
         }
+
+        public static ElementTheme ToElementTheme(this ApplicationTheme theme)
+        {
+            switch (theme)
+            {
+                case ApplicationTheme.Light:
+                    return ElementTheme.Light;
+                case ApplicationTheme.Dark:
+                default:
+                    return ElementTheme.Dark;
+            }
+        }
+
+        public static ApplicationTheme ToApplicationTheme(this ElementTheme theme)
+        {
+            switch (theme)
+            {
+                case ElementTheme.Default:
+                    return ApplicationTheme.Dark;
+                case ElementTheme.Light:
+                    return ApplicationTheme.Light;
+                case ElementTheme.Dark:
+                default:
+                    return ApplicationTheme.Dark;
+            }
+        }
     }
 }
